% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMfunctions.R
\name{convexADMM}
\alias{convexADMM}
\title{ADMM algorithm for Convex Cluster}
\usage{
convexADMM(X, gamma = 0.1, U = NULL, V = NULL, etainit = NULL, eps = 0.001)
}
\arguments{
\item{X}{A given n x p input matrix}

\item{gamma}{positive scalar, default value 0.1}

\item{U}{A given n x p cluster center matrix. If none is supplied, initialize with matrix of zeros.}

\item{V}{A given n x p cluster adjacency matrix(difference between the two adjacent centroids).If none is supplied, initialize with matrix of zeros.}

\item{etainit}{A n x p matrix, starting value of eta. If none is supplied, initialize with matrix of zeros.}

\item{eps}{A positive scalar, convergence tolerance criteria (difference in objective function values), default value is 0.001}
}
\value{
Returns a List with 2 members:

U - value of matrix U at convergence
V - value of matrix V at convergence
eta - value of matrix eta at convergence
}
\description{
ADMM algorithm for Convex Cluster
}
